ref: Move shared arroyo metrics wrapper
The MetricsWrapper is a compatibility layer for Arroyo that allows
all Arroyo instrumented metrics to be automatically sent to Sentry's
backend. This code was originally part of the `sentry_metrics` product
but is now moved to a shared utility as this can be reused to instrument
all Arroyo consumers going forward.
